Refactor and update dependencies
By adding exclude to the skydoc_repositories macro, we can allow the
caller to sort out dependencies so that they can resolve any particular
versioning needs for their internal use and between other libraries.
This prevents polluting the global //external space, which is
discouraged.

The outstanding problem here is protobuf, which expects some things
(notably six) to be available in //external, but for now I've just let
the caller deal with that as well if they so chose. I'll file a
follow-up ticket (and patch?) with protobuf to sort this out.